requirements.txt
fastapi
uvicorn
transformers
torch
newspaper3k
lxml_html_clean